refactor callback logic
