+++ app/models/address.rb
@@ -50,7 +50,7 @@ class Address
   before_save :set_crm_updates
   before_destroy :set_crm_updates
 
-  after_save :notify_address_changed, if: :changed?
+  after_update :notify_address_changed, if: :changed?
 
   track_history :on => [:fields],
                 :scope => :person,
@@ -341,12 +341,14 @@ class Address
   # @api private
   # @return [ job ID ]
   # @note This method is called by Mongoid after the document is saved.
-  #  It enqueues a job to notify the AddressWorker of the address change.
+  # calls the operation to build the payload and send it to the queue
+  # success or failure of the operation is not handled here or should not stop the save
   def notify_address_changed
     return unless EnrollRegistry.feature_enabled?(:notify_address_changed)
     return unless self.person.present?
 
-    AddressWorker.perform_async({address_id: self.id.to_s, person_hbx_id: self.person.hbx_id})
+    Operations::People::Addresses::AddressWorker.new.call({address_id: self.id.to_s, person_hbx_id: self.person.hbx_id})
+    true
   end
 
   def attribute_matches?(attribute, other)
